+++ src/main/java/kitchenpos/repository/ProductRepository.java
@@ -1,4 +1,4 @@
-package kitchenpos.application.repository;
+package kitchenpos.repository;
 
 import kitchenpos.domain.product.Product;
 import org.springframework.data.jpa.repository.JpaRepository;
